+++ link/src/main/java/com/stripe/android/link/ui/signup/SignUpScreen.kt
@@ -50,7 +50,11 @@ private fun SignUpBodyPreview() {
     DefaultLinkTheme {
         SignUpBody(
             merchantName = "Example, Inc.",
-            emailElement = EmailSpec.transform("email"),
+            emailElement = EmailSpec.transform(
+                mapOf(
+                    IdentifierSpec.Email to "email"
+                )
+            ),
             signUpState = SignUpState.InputtingPhone,
             onSignUpClick = {}
         )