+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/elements/AuBankAccountNumberSpec.kt
@@ -6,10 +6,19 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP_PREFIX)
 object AuBankAccountNumberSpec :
-    SectionFieldSpec(IdentifierSpec.Generic("au_becs_debit[account_number]")) {
-    fun transform(): SectionFieldElement =
+    SectionFieldSpec(
+        IdentifierSpec.Generic(
+            "au_becs_debit[account_number]"
+        )
+    ) {
+    fun transform(
+        initialValues: Map<IdentifierSpec, String?>
+    ): SectionFieldElement =
         SimpleTextElement(
             this.identifier,
-            SimpleTextFieldController(AuBankAccountNumberConfig())
+            SimpleTextFieldController(
+                AuBankAccountNumberConfig(),
+                initialValue = initialValues[this.identifier]
+            )
         )
 }