+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/elements/CardDetailsController.kt
@@ -4,22 +4,37 @@ import android.content.Context
 import kotlinx.coroutines.flow.combine
 import java.util.UUID
 
-internal class CardDetailsController constructor(context: Context) : SectionFieldErrorController {
+internal class CardDetailsController constructor(
+    context: Context,
+    initialValues: Map<IdentifierSpec, String?>
+) : SectionFieldErrorController {
 
     val label: Int? = null
     val numberElement = CardNumberElement(
         IdentifierSpec.CardNumber,
-        CardNumberController(CardNumberConfig(), context)
+        CardNumberController(
+            CardNumberConfig(),
+            context,
+            initialValue = initialValues[IdentifierSpec.CardNumber]
+        )
     )
 
     val cvcElement = CvcElement(
         IdentifierSpec.Generic("card[cvc]"),
-        CvcController(CvcConfig(), numberElement.controller.cardBrandFlow)
+        CvcController(
+            CvcConfig(),
+            numberElement.controller.cardBrandFlow,
+            initialValue = initialValues[IdentifierSpec.Generic("card[cvc]")]
+        )
     )
 
     val expirationDateElement = SimpleTextElement(
         IdentifierSpec.Generic("date"),
-        SimpleTextFieldController(DateConfig())
+        SimpleTextFieldController(
+            DateConfig(),
+            initialValue = initialValues[IdentifierSpec.Generic("card[exp_month]")] +
+                initialValues[IdentifierSpec.Generic("card[exp_year]")]?.takeLast(2)
+        ),
     )
 
     private val rowFields = listOf(expirationDateElement, cvcElement)