+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/elements/CardDetailsElement.kt
@@ -13,7 +13,8 @@ import kotlinx.coroutines.flow.combine
 internal class CardDetailsElement(
     identifier: IdentifierSpec,
     context: Context,
-    val controller: CardDetailsController = CardDetailsController(context),
+    initialValues: Map<IdentifierSpec, String?>,
+    val controller: CardDetailsController = CardDetailsController(context, initialValues),
 ) : SectionMultiFieldElement(identifier) {
     override fun sectionFieldErrorController(): SectionFieldErrorController =
         controller
@@ -43,7 +44,7 @@ internal class CardDetailsElement(
             val newString = convertTo4DigitDate(date)
             if (newString.length == 4) {
                 month = requireNotNull(newString.take(2).toIntOrNull())
-                year = requireNotNull(newString.takeLast(2).toIntOrNull())
+                year = requireNotNull(newString.takeLast(2).toIntOrNull()) + 2000
             }
         }