+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/elements/CardDetailsSpec.kt
@@ -7,7 +7,12 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP_PREFIX)
 object CardDetailsSpec : SectionFieldSpec(IdentifierSpec.Generic("card_details")) {
-    fun transform(context: Context): SectionFieldElement = CardDetailsElement(
-        IdentifierSpec.Generic("credit_detail"), context
+    fun transform(
+        context: Context,
+        initialValues: Map<IdentifierSpec, String?>
+    ): SectionFieldElement = CardDetailsElement(
+        IdentifierSpec.Generic("credit_detail"),
+        context,
+        initialValues
     )
 }