+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/elements/CardNumberController.kt
@@ -25,17 +25,20 @@ internal class CardNumberController constructor(
     cardAccountRangeRepository: CardAccountRangeRepository,
     workContext: CoroutineContext,
     staticCardAccountRanges: StaticCardAccountRanges = DefaultStaticCardAccountRanges(),
+    initialValue: String?,
     override val showOptionalLabel: Boolean = false
 ) : TextFieldController, SectionFieldErrorController {
 
     @JvmOverloads
     constructor(
         cardTextFieldConfig: CardNumberConfig,
-        context: Context
+        context: Context,
+        initialValue: String?
     ) : this(
         cardTextFieldConfig,
         DefaultCardAccountRangeRepositoryFactory(context).create(),
-        Dispatchers.IO
+        Dispatchers.IO,
+        initialValue = initialValue
     )
 
     override val capitalization: KeyboardCapitalization = cardTextFieldConfig.capitalization
@@ -54,7 +57,8 @@ internal class CardNumberController constructor(
     override val contentDescription: Flow<String> = _fieldValue
 
     internal val cardBrandFlow = _fieldValue.map {
-        accountRangeService.accountRange?.brand ?: CardBrand.getCardBrands(it).firstOrNull() ?: CardBrand.Unknown
+        accountRangeService.accountRange?.brand ?: CardBrand.getCardBrands(it).firstOrNull()
+            ?: CardBrand.Unknown
     }
 
     override val trailingIcon: Flow<TextFieldIcon?> = _fieldValue.map {
@@ -72,7 +76,9 @@ internal class CardNumberController constructor(
         cardTextFieldConfig.determineState(
             brand,
             fieldValue,
-            accountRangeService.accountRange?.panLength ?: brand.getMaxLengthForCardNumber(fieldValue)
+            accountRangeService.accountRange?.panLength ?: brand.getMaxLengthForCardNumber(
+                fieldValue
+            )
         )
     }
     override val fieldState: Flow<TextFieldState> = _fieldState
@@ -87,7 +93,8 @@ internal class CardNumberController constructor(
         object : CardAccountRangeService.AccountRangeResultListener {
             override fun onAccountRangeResult(newAccountRange: AccountRange?) {
                 newAccountRange?.panLength?.let { panLength ->
-                    (visualTransformation as CardNumberVisualTransformation).binBasedMaxPan = panLength
+                    (visualTransformation as CardNumberVisualTransformation).binBasedMaxPan =
+                        panLength
                 }
             }
         }
@@ -116,7 +123,7 @@ internal class CardNumberController constructor(
         }
 
     init {
-        onValueChange("")
+        onRawValueChange(initialValue ?: "")
     }
 
     /**