+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/elements/IbanSpec.kt
@@ -6,9 +6,12 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP_PREFIX)
 object IbanSpec : SectionFieldSpec(IdentifierSpec.Generic("sepa_debit[iban]")) {
-    fun transform(): SectionFieldElement =
+    fun transform(initialValues: Map<IdentifierSpec, String?>): SectionFieldElement =
         IbanElement(
             this.identifier,
-            SimpleTextFieldController(IbanConfig())
+            SimpleTextFieldController(
+                IbanConfig(),
+                initialValue = initialValues[this.identifier]
+            )
         )
 }