+++ payments-ui-core/src/test/java/com/stripe/android/ui/core/forms/TransformSpecToElementTest.kt
@@ -71,7 +71,6 @@ internal class TransformSpecToElementTest {
                 ),
                 initialValues = mapOf(),
                 amount = null,
-                country = "DE",
                 saveForFutureUseInitialValue = true,
                 merchantName = "Merchant, Inc.",
                 context