+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/BaseAddPaymentMethodFragment.kt
@@ -20,6 +20,7 @@ import androidx.lifecycle.lifecycleScope
 import com.stripe.android.core.injection.InjectorKey
 import com.stripe.android.model.CardBrand
 import com.stripe.android.model.PaymentMethod
+import com.stripe.android.model.PaymentMethodCreateParams
 import com.stripe.android.model.StripeIntent
 import com.stripe.android.paymentsheet.databinding.FragmentPaymentsheetAddPaymentMethodBinding
 import com.stripe.android.paymentsheet.forms.FormFieldValues
@@ -106,6 +107,9 @@ internal abstract class BaseAddPaymentMethodFragment : Fragment() {
             val formViewModel = formFragment.formViewModel
             viewLifecycleOwner.lifecycleScope.launch {
                 formViewModel.completeFormValues.collect { formFieldValues ->
+                    // if the formFieldValues is a change either null or new values for the
+                    // newLpm then we should clear it out --- but what happens if we cancel -- selection should
+                    // have the correct value
                     sheetViewModel.updateSelection(
                         transformToPaymentSelection(
                             formFieldValues,
@@ -165,7 +169,27 @@ internal abstract class BaseAddPaymentMethodFragment : Fragment() {
                 showPaymentMethod = paymentMethod,
                 merchantName = sheetViewModel.merchantName,
                 amount = sheetViewModel.amount.value,
-                injectorKey = sheetViewModel.injectorKey
+                injectorKey = sheetViewModel.injectorKey,
+                initialPaymentMethodCreateParams =
+                if (sheetViewModel.newLpm?.paymentMethodCreateParams?.typeCode ==
+                    paymentMethod.type.code
+                ) {
+                    when (sheetViewModel.newLpm) {
+                        is PaymentSelection.New.GenericPaymentMethod -> {
+                            (sheetViewModel.newLpm as PaymentSelection.New.GenericPaymentMethod)
+                                .paymentMethodCreateParams
+                        }
+                        is PaymentSelection.New.Card -> {
+                            (sheetViewModel.newLpm as PaymentSelection.New.Card)
+                                .paymentMethodCreateParams
+                        }
+                        else -> {
+                            null
+                        }
+                    }
+                } else {
+                    null
+                }
             )
         )
 
@@ -245,7 +269,8 @@ internal abstract class BaseAddPaymentMethodFragment : Fragment() {
             config: PaymentSheet.Configuration?,
             merchantName: String,
             amount: Amount? = null,
-            @InjectorKey injectorKey: String
+            @InjectorKey injectorKey: String,
+            initialPaymentMethodCreateParams: PaymentMethodCreateParams?
         ): FormFragmentArguments {
 
             val layoutFormDescriptor = showPaymentMethod.getPMAddForm(stripeIntent, config)
@@ -257,7 +282,8 @@ internal abstract class BaseAddPaymentMethodFragment : Fragment() {
                 merchantName = merchantName,
                 amount = amount,
                 billingDetails = config?.defaultBillingDetails,
-                injectorKey = injectorKey
+                injectorKey = injectorKey,
+                initialPaymentMethodCreateParams = initialPaymentMethodCreateParams
             )
         }
     }