+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/viewmodels/BaseSheetViewModel.kt
@@ -36,6 +36,7 @@ import com.stripe.android.paymentsheet.model.FragmentConfig
 import com.stripe.android.paymentsheet.model.PaymentSelection
 import com.stripe.android.paymentsheet.model.SavedSelection
 import com.stripe.android.paymentsheet.model.SupportedPaymentMethod
+import com.stripe.android.paymentsheet.paymentdatacollection.ComposeFormDataCollectionFragment
 import com.stripe.android.paymentsheet.repositories.CustomerRepository
 import com.stripe.android.ui.core.Amount
 import com.stripe.android.ui.core.forms.resources.ResourceRepository
@@ -48,7 +49,7 @@ import org.jetbrains.annotations.TestOnly
 import kotlin.coroutines.CoroutineContext
 
 @VisibleForTesting
-class TransitionFragmentResource {
+class MultiStepContinueIdlingResource {
     companion object {
         @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP_PREFIX)
         @Nullable
@@ -155,7 +156,7 @@ internal abstract class BaseSheetViewModel<TransitionTargetType>(
     internal val liveMode: LiveData<Boolean> = _liveMode
 
     /**
-     * On [CardDataCollectionFragment] this is set every time the details in the add
+     * On [ComposeFormDataCollectionFragment] this is set every time the details in the add
      * card fragment is determined to be valid (not necessarily selected)
      * On [BasePaymentMethodsListFragment] this is set when a user selects one of the options
      */
@@ -180,7 +181,7 @@ internal abstract class BaseSheetViewModel<TransitionTargetType>(
      * and reopen the card view. It is used on the Payment Options sheet similar to what is
      * described above, and when you have an unsaved card.
      */
-    abstract var newCard: PaymentSelection.New.Card?
+    abstract var newLpm: PaymentSelection.New?
 
     abstract fun onFatal(throwable: Throwable)
 
@@ -209,7 +210,6 @@ internal abstract class BaseSheetViewModel<TransitionTargetType>(
     }.distinctUntilChanged()
 
     init {
-        TransitionFragmentResource.idlingResource?.increment()
         if (_savedSelection.value == null) {
             viewModelScope.launch {
                 val savedSelection = withContext(workContext) {
@@ -273,9 +273,6 @@ internal abstract class BaseSheetViewModel<TransitionTargetType>(
     }
 
     open fun transitionTo(target: TransitionTargetType) {
-        if (TransitionFragmentResource.idlingResource?.isIdleNow == false) {
-            TransitionFragmentResource.idlingResource?.decrement()
-        }
         _transition.postValue(Event(target))
     }
 
@@ -337,6 +334,9 @@ internal abstract class BaseSheetViewModel<TransitionTargetType>(
     }
 
     fun updateSelection(selection: PaymentSelection?) {
+        if (selection is PaymentSelection.New) {
+            newLpm = selection
+        }
         savedStateHandle[SAVE_SELECTION] = selection
     }
 
@@ -345,9 +345,11 @@ internal abstract class BaseSheetViewModel<TransitionTargetType>(
     }
 
     fun getAddFragmentSelectedLpm() =
-        savedStateHandle.getLiveData<SupportedPaymentMethod>(
+        savedStateHandle.getLiveData(
             SAVE_SELECTED_ADD_LPM,
-            SupportedPaymentMethod.Card
+            SupportedPaymentMethod.fromCode(
+                newLpm?.paymentMethodCreateParams?.typeCode
+            ) ?: SupportedPaymentMethod.Card
         )
 
     fun getAddFragmentSelectedLpmValue() =
@@ -376,7 +378,7 @@ internal abstract class BaseSheetViewModel<TransitionTargetType>(
         }
     }
 
-    protected fun setupLink(stripeIntent: StripeIntent) {
+    protected fun setupLink(unused: StripeIntent) {
         // TODO(brnunes-stripe): Enable Link
         _isLinkEnabled.value = false
     }